Create Check Current Projects Spreadsheet
Looks every 30 minutes at the current project spreadsheet and updates which projects are available for signups. If a project's signup period has ended, it emails the project chairperson with a form for them to document how many hours each volunteer served for. That form will automatically update the member's totals on the "Total Service Hours" Spreadsheet.