Create Check Project Signups
Because google scripts is not always reliably fast, sometimes a person's signup response can be overwritten on the spreadsheet. To avoid this problem, I have the script pull responses from the "Responses", and then gradually install them into the spreadsheet, with checks to make sure that the entry was properly recorded. See the script for more details.