Create Submit an activity
This responds to a member of the Circle K board creating a project. It parses the information to make sure that it is all logical, then it automatically puts the project onto the "current projects" form for members to sign up for.

In progress: changing this script to a time based one to avoid losing information due to overwrite errors.

Note: The Google URL shortener API is currently functional, but is not in use to protect the privacy of the individual members. Am currently working on a fix to display the project roster without leaking members' email and timestamp information.